Wrap function bodies in `when` conditions
This reverses checking for when before defining the functions to
instead putting the when condition inside the function body. I
grepped around the el files that come with emacs and this appears
to be the standard form.

This also fixes some of the function comments to include the
argument names like they should.